combination of 39 commits, squashed

+++ workloads/cyclictest/common.sh
@@ -1,4 +1,5 @@
 source env.sh
+source ../../utils/benchmark-operator.sh
 
 # If ES_SERVER is set and empty we disable ES indexing and metadata collection
 if [[ -v ES_SERVER ]] && [[ -z ${ES_SERVER} ]]; then
@@ -71,6 +72,21 @@ export_defaults() {
   else
     echo $cloud_name > uuid.txt
   fi
+
+  #Check to see if the infrastructure type is baremetal to adjust script as necessary 
+  if [[ "${baremetalCheck}" == '"BareMetal"' ]]; then
+    log "BareMetal infastructure: setting isBareMetal accordingly"
+    export isBareMetal=true
+  else
+    export isBareMetal=false
+  fi
+
+  if [[ "${isBareMetal}" == "true" ]]; then
+    #Installing python3.8
+    sudo yum -y install python3.8
+    sudo alternatives --set python3 /usry/bin/python3.8
+  fi
+
 } 
 
 
@@ -146,29 +162,24 @@ deploy_perf_profile() {
 }
 
 deploy_operator() {
-  if [[ "${isBareMetal}" == "false" ]]; then
-    log "Removing benchmark-operator namespace, if it already exists"
-    oc delete namespace benchmark-operator --ignore-not-found
-    log "Cloning benchmark-operator from branch ${operator_branch} of ${operator_repo}"
-  else
-    log "Baremetal infrastructure: Keeping benchmark-operator namespace"
-    log "Cloning benchmark-operator from branch ${operator_branch} of ${operator_repo}"
+  deploy_benchmark_operator ${OPERATOR_REPO} ${OPERATOR_BRANCH}
+  if [[ $? != 0 ]]; then
+     exit 1
   fi
-    rm -rf benchmark-operator  
-    git clone --single-branch --branch ${operator_branch} ${operator_repo} --depth 1
-    (cd benchmark-operator && make deploy)
-    oc wait --for=condition=available "deployment/benchmark-controller-manager" -n benchmark-operator --timeout=300s
-    oc adm policy -n benchmark-operator add-scc-to-user privileged -z benchmark-operator
-    oc adm policy -n benchmark-operator add-scc-to-user privileged -z backpack-view
-    oc patch scc restricted --type=merge -p '{"allowHostNetwork": true}'
+  rm -rf benchmark-operator
 }
 
-
-deploy_workload() {
-  log "Deploying cyclictest benchmark"
-  envsubst < $CRD | oc apply -f -
-  log "Sleeping for 60 seconds"
-  sleep 60
+run_workload() {
+  log "Deploying benchmark"
+  local TMPCR=$(mktemp)
+  envsubst < $1 > ${TMPCR}
+  run_benchmark ${TMPCR} ${TEST_TIMEOUT}
+  local rc=$?
+  if [[ ${TEST_CLEANUP} == "true" ]]; then
+    log "Cleaning up benchmark"
+    kubectl delete -f ${TMPCR}
+  fi
+  return ${rc}
 }
 
 check_logs_for_errors() {
@@ -232,19 +243,6 @@ generate_csv() {
   # tbd
 }
 
-init_cleanup() {
-  if [[ "${isBareMetal}" == "false" ]]; then
-    log "Cloning benchmark-operator from branch ${operator_branch} of ${operator_repo}"
-    rm -rf /tmp/benchmark-operator
-    git clone --single-branch --branch ${operator_branch} ${operator_repo} /tmp/benchmark-operator --depth 1
-    oc delete -f /tmp/benchmark-operator/deploy
-    oc delete -f /tmp/benchmark-operator/resources/crds/ripsaw_v1alpha1_ripsaw_crd.yaml
-    oc delete -f /tmp/benchmark-operator/resources/operator.yaml
-  else
-    log "BareMetal Infrastructure: Skipping cleanup"
-  fi
-}
-
 delete_benchmark() {
   oc delete benchmarks.ripsaw.cloudbulldozer.io/cyclictest -n benchmark-operator
 }
@@ -267,9 +265,8 @@ normal=$(tput sgr0)
 python3 -m pip install -r requirements.txt | grep -v 'already satisfied'
 check_cluster_present
 export_defaults
-init_cleanup
 check_cluster_health
 deploy_perf_profile
 deploy_operator
-deploy_workload
-wait_for_benchmark
+run_workload ripsaw-cyclictest-crd.yaml
+#wait_for_benchmark
